adding redux. creating indexCrad and singleArticleCard
